update `updateNotification`
Wrapped up this one. A tiny component that required a massive rework to useInteractiveText.

useInteractiveText → interactiveTextStyle. This can now be used both for basic interactive text, or interactive contained text elements.

As part of this update I refactored a bunch of the intensity code, including changing ElementIntensities to the more descriptive and hopefully understandable IntensitySet.

Intensities that haven't been resolved yet from the theme have been changed to UnresolvedIntensity and UnresolvedIntensitySet. Hopefully this will make it obvious if you are using the wrong one, since intensities should only need to be resolved when building the theme.